+++ stac_fastapi/elasticsearch/stac_fastapi/elasticsearch/app.py
@@ -1,14 +1,26 @@
 """FastAPI application."""
+import os
+
 from stac_fastapi.api.app import StacApi
 from stac_fastapi.api.models import create_get_request_model, create_post_request_model
-from stac_fastapi.elasticsearch.config import ElasticsearchSettings
 from stac_fastapi.elasticsearch.core import (
     BulkTransactionsClient,
     CoreClient,
     EsAsyncBaseFiltersClient,
     TransactionsClient,
 )
-from stac_fastapi.elasticsearch.database_logic import create_collection_index
+
+if os.getenv("BACKEND", "elasticsearch").lower() == "opensearch":
+    from stac_fastapi.elasticsearch.config.config_opensearch import SearchSettings
+    from stac_fastapi.elasticsearch.database_logic.database_logic_opensearch import (
+        create_collection_index,
+    )
+else:
+    from stac_fastapi.elasticsearch.config.config_elasticsearch import SearchSettings
+    from stac_fastapi.elasticsearch.database_logic.database_logic_elasticsearch import (
+        create_collection_index,
+    )
+
 from stac_fastapi.elasticsearch.extensions import QueryExtension
 from stac_fastapi.elasticsearch.session import Session
 from stac_fastapi.extensions.core import (
@@ -21,7 +33,7 @@ from stac_fastapi.extensions.core import (
 )
 from stac_fastapi.extensions.third_party import BulkTransactionExtension
 
-settings = ElasticsearchSettings()
+settings = SearchSettings()
 session = Session.create_from_settings(settings)
 
 filter_extension = FilterExtension(client=EsAsyncBaseFiltersClient())