+++ stac_fastapi/elasticsearch/stac_fastapi/elasticsearch/config/config_opensearch.py
@@ -3,7 +3,8 @@ import os
 import ssl
 from typing import Any, Dict, Set
 
-from elasticsearch import AsyncElasticsearch, Elasticsearch  # type: ignore
+from opensearchpy import AsyncOpenSearch, OpenSearch
+
 from stac_fastapi.types.config import ApiSettings
 
 
@@ -18,7 +19,7 @@ def _es_config() -> Dict[str, Any]:
     # Initialize the configuration dictionary
     config = {
         "hosts": hosts,
-        "headers": {"accept": "application/vnd.elasticsearch+json; compatible-with=7"},
+        "headers": {"accept": "application/json", "Content-Type": "application/json"},
     }
 
     # Explicitly exclude SSL settings when not using SSL
@@ -39,13 +40,22 @@ def _es_config() -> Dict[str, Any]:
     if (u := os.getenv("ES_USER")) and (p := os.getenv("ES_PASS")):
         config["http_auth"] = (u, p)
 
+    if api_key := os.getenv("ES_API_KEY"):
+        if isinstance(config["headers"], dict):
+            headers = {**config["headers"], "x-api-key": api_key}
+
+        else:
+            config["headers"] = {"x-api-key": api_key}
+
+        config["headers"] = headers
+
     return config
 
 
 _forbidden_fields: Set[str] = {"type"}
 
 
-class ElasticsearchSettings(ApiSettings):
+class SearchSettings(ApiSettings):
     """API settings."""
 
     # Fields which are defined by STAC but not included in the database model
@@ -55,10 +65,10 @@ class ElasticsearchSettings(ApiSettings):
     @property
     def create_client(self):
         """Create es client."""
-        return Elasticsearch(**_es_config())
+        return OpenSearch(**_es_config())
 
 
-class AsyncElasticsearchSettings(ApiSettings):
+class AsyncSearchSettings(ApiSettings):
     """API settings."""
 
     # Fields which are defined by STAC but not included in the database model
@@ -68,4 +78,4 @@ class AsyncElasticsearchSettings(ApiSettings):
     @property
     def create_client(self):
         """Create async elasticsearch client."""
-        return AsyncElasticsearch(**_es_config())
+        return AsyncOpenSearch(**_es_config())